+++ src/runtime/Converter.cs
@@ -133,7 +133,8 @@ namespace Python.Runtime
             if (EncodableByUser(type, value))
             {
                 var encoded = PyObjectConversions.TryEncode(value, type);
-                if (encoded != null) {
+                if (encoded != null)
+                {
                     return new NewReference(encoded);
                 }
             }
@@ -334,7 +335,7 @@ namespace Python.Runtime
 
             if (obType.IsGenericType && obType.GetGenericTypeDefinition() == typeof(Nullable<>))
             {
-                if( value == Runtime.PyNone )
+                if (value == Runtime.PyNone)
                 {
                     result = null;
                     return true;
@@ -686,10 +687,8 @@ namespace Python.Runtime
                         {
                             if (Runtime.PyUnicode_GetLength(value) == 1)
                             {
-                                IntPtr unicodePtr = Runtime.PyUnicode_AsUnicode(value);
-                                Char[] buff = new Char[1];
-                                Marshal.Copy(unicodePtr, buff, 0, 1);
-                                result = buff[0];
+                                int chr = Runtime.PyUnicode_ReadChar(value, 0);
+                                result = (Char)chr;
                                 return true;
                             }
                             goto type_error;
@@ -981,6 +980,12 @@ namespace Python.Runtime
         {
             if (o is null) return Runtime.None;
             return Converter.ToPython(o, o.GetType()).MoveToPyObject();
+        }
+
+        public static PyObject ToPythonAs<T>(this T? o)
+        {
+            if (o is null) return Runtime.None;
+            return Converter.ToPython(o, typeof(T)).MoveToPyObject();
         }
     }
 }